From David Bolen:
You may or may not have seen the brief discussion on c.l.py regarding
handling exit codes for the higher order os.popen#sf (#>=2) functions, but in
response to that I've created an augmented version of my patch for exit
codes.  What it adds is tracking all file handles returned for any
os.popen* call, and it only waits for the child and retrieves the exit code
when the last handle is closed.  This should prevent any potential deadlock
situation where for some cases the handles needed to be closed such that the
child stdin wasn't last (which the prior patch always used to wait on).

The patch also touches the win32popenWin9x.cpp file, since that module was
not currently bubbling up the result codes from its child processes, so
without a small change, this stuff wouldn't have worked under Win95/98
anyway.