setup.py: We are looking for headers not sources
The comment mentions, that we are looking for the MFC source code. I'm not an expert, but I doubt that Microsoft is sharing the source code of MFC on their Visual Studio releases. Therefore I'm correcting the comment to what we are looking for in the following code, that we are looking for headers.