Fixes Traceback when using Sapi.SpVoice
A traceback occurs when setting the Voice property in  Sapi.SpVoice

code to reproduce the problem.

    import win32com.client
    import pythoncom
    
    pythoncom.CoInitialize()
    tts = win32com.client.Dispatch("Sapi.SpVoice")
    tts.Voice = tts.GetVoices('Name=Microsoft Anna')[0]

The traceback

    Traceback (most recent call last):
      File "", line 6, in <module>
        tts.Voice = tts.GetVoices('Name=Microsoft Anna')[0]
      File "\win32com\client\dynamic.py", line 565, in __setattr__
        self._oleobj_.Invoke(entry.dispid, 0, invoke_type, 0, value)
    pywintypes.com_error: (-2147352573, 'Member not found.', None, None)

I have not come across anything that this fix causes an issue with. Because of the broad scope of com I am not able to test everything. This may give some insight on where the problem may lie.