Make PyIDirectorySearch work on 64-bit: prep
I'm going to need to change the ParseTuple calls in some of the pure-SWIG methods in PyIDirectorySearch.

With my limited-to-none experience I can't see an obvious way to fix the mapping in SWIG itself, so break out each of the affected methods into native code (like is done for the other methods in this class).