Feature #1563: Report SERVICE_STOPPED with an error code if SvcRun() raises an Exception.
If SvcRun() raises an Exception then the final call to SetServiceStatus will be with a SERVICE_STATUS structure,
with dwCurrentState=SERVICE_STOPPED, dwWin32ExitCode=ERROR_SERVICE_SPECIFIC_ERROR and dwServiceSpecificExitCode non-zero.

This is so Services that have the "Enable Actions For Stops With Errors" / `failureflag` option enabled,
will have their recovery actions run when SvcRun() raises an Exception.