Simplify pywin32_bootstrap, avoid importing site
Since pywin32 has dropped support for Python 2.7 since version 300 and support
for Python 3.0 through 3.4 since version 223, we can now rely on [PEP-420]
semantics when importing names of directories available on `sys.path` and not
containing an `__init__.py`. This applies to to `pywin32_system32`. The previous
version of `win32.lib.pywin32_boostrap` manually searched site-packages
directories, stopping at the first one found. Now we pass this responsibility
off to Python's import machinery. When we `import pywin32_system32`, if
successful, `pywin32_system32` will likely be a
`_frozen_importlib_external._NamespacePath` object. We obtain its first entry,
being careful that it didn't support sequence indexing until Python 3.7 or so
(and in any case, that's all undocumented; the [documentation promises] only
that `__path__: Iterable[str]`). This first entry is the path to the directory
containing the DLLs, just as before, so the remainder of the code is unchanged.

The only semantic differences between this implementation and the previous one
is that
1. we no longer import `site` (always better to import less!), and
2. `pywin32_system32` is no longer _required_ to be in a site-packages directory

Just as with the previous implementation, this one does not raise an exception
if `pywin32_system32` cannot be found.

I did one annoying thing in this patch to minimize the diff, which is that I
redefine the variable `pywin32_system32`: first it refers to a namespace package
module object, and then later it refers to the `str` directory name. It might be
reasonable after initial review of the patch to rename the latter instance of
the variable to something like `path`.

Finally, I added no tests and no change log entry, but can if necessary.

[PEP-420]: https://www.python.org/dev/peps/pep-0420/#specification
[documentation promises]: https://docs.python.org/3/reference/import.html#__path__