Fix PythonService.exe initialization when starting as NT_AUTHORITY/SYSTEM
Since the advent of Python3, the python windows installer now install
python "as a user" (in the LocalFolder) and optionnally adds sys.prefix
to the **user's PATH**.

However, PythonService.exe by default runs as NT_AUTHORITY/SYSTEM which
does not share the same %PATH% environment variables (in order to
prevent privilege escalations paths based on dll hijacking). Currently,
on a fresh install, PythonService.exe can not launch since it rely on
pythonXXX.dll which is not present on SYSTEM's path.

This commit aims to fix this situation by ensuring PythonXXX.dll is in
%System32%, a folder both on the DLL loading path for any users or
SYSTEM.