DDE client request in either Unicode or ASCII in Python 3
DDE client request call in Python 3 can fail if the call is made to a
DDE server that can only serve text in ASCII (CF_TEXT).

This change modifies the request call such that if the request call in
Unicode (CF_UNICODETEXT) fails then the request call is made in ASCII
instead.

This issue was notified when making a request call to the service "Excel",
topic "System", item "Topics".

This commit may fix this issue:

    https://github.com/mhammond/pywin32/issues/1241