Various ssize_t cleanups.
Primarily just avoids warning when converting between Py_ssize_t and ints.
Many cases the sizes are checked, but other cases just blindly cast - the
latter is done when it seems almost impossible the int will be large
enough to overflow, but that choice might seem arbitrary in some cases.