Simplify and improve pywintypes search for dll in site-packages.
The current method only looks at "global" and "user" site,
which is not enough if pywin32 is installed in "system" site packages
and imported from a venv created with --system-site-packages.

With this updated approach, we avoid looking at each site separately
and let the import statement find the location of pywin32_system32.