Handle `varchar(max)`/`nvarchar(max)` column sizes
`SQLDescribeCol()` returns 0 ("unlimited") as column sizes for `varchar(max)`/`nvarchar(max)`/`varbinary(max)` types. We use the cursor's maximum output size instead in this case.

The change for `SQL_BINARY`/`SQL_VARBINARY` cases is not a must but fits the idea of the original commit (c91a0f1).

Fixes mhammond#1953.