Fix to make special __dunder__ methods work with CoClass.
This was initially implemented in commit
17fb589d614e998bc3188f9e6200beb13de772eb to address #1699.

The initial implementation had a problem with bool() (see #1753).
The problem was caused by implementing __nonzero__ instead of __bool__,
which was pointed out in the conclusion of #1870.
The commit c072ff7679fd52003f32596be1d2620c1189f8d8 tried to fix #1753
but introduced another issue #1870.