Integrate PyCOMTest server dll into CI workflow (#2493)
* Added Visual Studio 2022 solution/project files for PyCOMTest server dll.

Also changed '.gitignore' to put the Visual Studio solution
and pjoject files under revision control.

* Added building and registering the
PyCOMTest server dll to the CI workflow.

* Retarget PyCOMTest solution for MSVC v142 and Win SDK 10.0.22000.0

* Moved PyCOMTest to 'Level 1' of the unit tests.

This will include PyCOMTest in the tests of the github CI workflow.

* Added registration of the PyCOMTest.dll with regsvr32 to the CI workflow.

* Fixing the registred trademark symbol in'PyCOMTest.idl'.

* Fix to generate a standalone PyCOMTestLib module and not a package.

* Add bForDemand=False to testArrays.py

* Fix formatting.

* Fix to make special __dunder__ methods work with CoClass.

This was initially implemented in commit
17fb589d614e998bc3188f9e6200beb13de772eb to address #1699.

The initial implementation had a problem with bool() (see #1753).
The problem was caused by implementing __nonzero__ instead of __bool__,
which was pointed out in the conclusion of #1870.
The commit c072ff7679fd52003f32596be1d2620c1189f8d8 tried to fix #1753
but introduced another issue #1870.

* Added a change entry for the #1870 fix.

---------

Co-authored-by: Avasam <samuel.06@hotmail.com>