Defer creating package in makepy when bForDemand is set (#2678)
* Update testmakepy.py to test concurrent calls to EnsureModule

This test currently fails because of issue #1923

* Defer creating package in makepy when bForDemand is set

Creating the folder without the mutex acquired means that other
processes can import it as a namespace package before the generation is
finished.

This change generates the temp file in the gen_py folder without
creating the package until the writer is finished and the mutex is held.

Fixes #1923
Co-authored-by: Avasam <samuel.06@hotmail.com>